a lot done in back end

+++ src/main/kotlin/com/aris/booklibraries/demoBookLibraries/executors/BookExecutor.kt
@@ -2,9 +2,7 @@ package com.aris.booklibraries.demoBookLibraries.executors
 
 import com.aris.booklibraries.demoBookLibraries.models.Book
 import com.aris.booklibraries.demoBookLibraries.models.response.ApiResponse
-import com.aris.booklibraries.demoBookLibraries.services.AuthorService
 import com.aris.booklibraries.demoBookLibraries.services.BookService
-import com.aris.booklibraries.demoBookLibraries.services.LibraryService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Component
@@ -14,29 +12,20 @@ import javax.servlet.http.HttpServletResponse
 class BookExecutor {
     @Autowired
     lateinit var bookService: BookService
-    @Autowired
-    lateinit var authorService: AuthorService
-    @Autowired
-    lateinit var libraryService: LibraryService
 
     fun getAllBooks(): ApiResponse< List<Book>,String> {
         val allBooks = bookService.findAll()
-
-        return if (allBooks != null) {
-            ApiResponse(data = allBooks)
-        } else {
-            ApiResponse(null,message = "Something went wrong")
-        }
+        return ApiResponse(data = allBooks, message = "OK")
     }
 
     fun getBookById(bookId: Long,response: HttpServletResponse): ApiResponse<Book,String> {
         val bookToReturn = bookService.findById(bookId)
 
         return if ( bookToReturn != null ) {
-            ApiResponse(data=bookToReturn)
+            ApiResponse(data=bookToReturn, message = "OK")
         } else {
             response.status = HttpStatus.BAD_REQUEST.value()
-            ApiResponse(data = null, message = "No book with such id.")
+            ApiResponse(data = null, message = "Error: No book with such id.")
         }
     }
 
@@ -48,30 +37,22 @@ class BookExecutor {
         }
 
         val bookData = Book(null, data.title, null)
-        if (data.author != null) {
+        return if (data.author != null) {
             val createdBook = bookService.addBook(bookData, data.author!!)
             response.status = HttpStatus.ACCEPTED.value()
-            return ApiResponse(data = createdBook)
+            ApiResponse(data = createdBook, message = "OK")
 
         } else {
             response.status = HttpStatus.BAD_REQUEST.value()
-            return ApiResponse(data = null, message = "No author found")
+            ApiResponse(data = null, message = "No author found")
         }
     }
 
-    fun deleteById(bookId: Long) {
-        libraryService?.deleteBookById(bookId)
-        bookService.deleteById(bookId)
+    fun deleteById(bookId: Long, response: HttpServletResponse): ApiResponse<String,String> {
+        val bookToDelete = bookService.findById(bookId)
+        if ( bookToDelete != null) {
+            bookService.deleteById(bookId)
+        }
+        return ApiResponse(data = null, message = "OK")
     }
-
-    fun deleteByAuthor(authorId: Long) {
-        //get the id of the books with author id
-        //delete them from has_books
-        //then finally delete book
-
-
-        bookService.deleteByAuthor(authorId)
-    }
-
-
 }
